fixes pequenos e modularização dos dialogs #18
