+++ src/app/app.module.ts
@@ -23,8 +23,11 @@ import {MatSelectModule} from '@angular/material/select';
 // Custom modules
 import {CardapioComponent} from './cardapio/cardapio.component';
 import { ItemCardapioComponent } from './item-cardapio/item-cardapio.component';
-import { CardapioBotoesComponent, ExcluirItemDialog, NovoItemDialog, NovaCategoriaDialog } from './cardapio-botoes/cardapio-botoes.component'
+import { CardapioBotoesComponent } from './cardapio-botoes/cardapio-botoes.component'
 import { BasicSnackBar } from './basic.snackbar';
+import { NovoItemDialog } from './cardapio-botoes/novo-item-dialog'
+import { ExcluirItemDialog } from './cardapio-botoes/excluir-item-dialog'
+import { NovaCategoriaDialog } from './cardapio-botoes/nova-categoria-dialog'
 
 @NgModule({
   declarations: [
