+++ src/app/cardapio-botoes/excluir-item-dialog.ts
@@ -1,0 +1,8 @@
+import { Component, OnInit } from '@angular/core';
+
+@Component({
+    selector: 'excluir-item-dialog',
+    templateUrl: 'excluir-item-dialog.html'
+})
+
+export class ExcluirItemDialog {}
