+++ src/app/cardapio-botoes/nova-categoria-dialog.html
@@ -9,7 +9,7 @@
 </div>
 
 <div mat-dialog-actions>
-    <button mat-raised-button (click)="sendCategoria()">Excluir Item</button>
+    <button mat-raised-button (click)="sendCategoria()">Criar Categoria</button>
     <button mat-raised-button mat-dialog-close>Fechar</button>
     <mat-error *ngIf="invalidForm"> A categoria já existe! </mat-error>
